fix: 完成响应质量全面审计修复，提升数据一致性与准确性
- 进行了系统扫描，修复了 11 项关键问题，包括比特率类型更改、容器格式规范化、最大比特率比较修复等，确保数据在 DTO 构建和流媒体端点中的一致性。
- 新增了动态推导播放方法，优化了会话状态的处理，确保更准确的播放信息。
- 更新了兼容性报告，记录此次审计修复的细节与影响。

+++ backend/src/models.rs
@@ -1226,7 +1226,7 @@ pub struct BaseItemDto {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub file_name: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub bitrate: Option<i32>,
+    pub bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub official_rating: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -1418,7 +1418,7 @@ pub struct MediaSourceDto {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub e_tag: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub bitrate: Option<i32>,
+    pub bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub default_audio_stream_index: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -1598,11 +1598,11 @@ pub struct TranscodingInfoDto {
     pub is_video_direct: bool,
     pub is_audio_direct: bool,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub bitrate: Option<i32>,
+    pub bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub audio_bitrate: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub video_bitrate: Option<i32>,
+    pub video_bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub framerate: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
